+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Drag-and-drop support: drag files from project view to hot folders
+- Drag-and-drop support: drag files from hot folders to project (creates a copy)
+- Automatic filtering of .DS_Store files from hot folder tree view
+
+### Changed
+- Improved file visibility by hiding system files (.DS_Store) automatically
+
 ## [1.0.0] - 2026-01-09
 
 ### Added